fixed search action bugs

+++ TextWarp/Controllers/GalleryController.cs
@@ -4,6 +4,7 @@ using TextWarp.Models.Database;
 using TextWarp.Models;
 using Microsoft.AspNetCore.Authorization;
 using System.Security.Policy;
+using System.Linq;
 
 namespace TextWarp.Controllers
 {
@@ -21,26 +22,35 @@ namespace TextWarp.Controllers
         }
 
         [Route("gallery/getData/")]
-        [HttpGet]
-        public ActionResult getData()
+        [HttpPost]
+        public ActionResult getData(WorkSearchModel workSearchItem)
         {
             try
             {
-                var saved_svgs = _context.WarpedSvgs.Where(s => s.UserId == "41ae9ea6-035a-4bc6-98f9-fd758422de6d").ToList();
-                if (saved_svgs != null)
+                if(workSearchItem.WorkName != null)
                 {
-                    foreach (var saved_svg in saved_svgs)
-                    {
-                        if (saved_svg.SvgfileName == "")
-                        {
-                            _context.WarpedSvgs.Remove(saved_svg);
-                            _context.SaveChanges();
-                        }
-                    }
-                    saved_svgs = _context.WarpedSvgs.Where(s => s.UserId == "41ae9ea6-035a-4bc6-98f9-fd758422de6d" && s.SvgfileName != "").OrderByDescending(s => s.UpdatedAt).ToList();
-                    return Json(new {status = "success", saved_svgs = saved_svgs});
+                    var saved_svgs = _context.WarpedSvgs.Where(s => s.UserId == "41ae9ea6-035a-4bc6-98f9-fd758422de6d" && s.WorkName.Contains(workSearchItem.WorkName)).ToList();
+                    if (saved_svgs != null) return Json(new { status = "success", saved_svgs = saved_svgs });
+                    return Json(new { status = "success", msg = "There is no svgs." });
                 }
-                return Json(new {status = "success", msg = "There is no svgs."});
+                else
+                {
+                    var saved_svgs = _context.WarpedSvgs.Where(s => s.UserId == "41ae9ea6-035a-4bc6-98f9-fd758422de6d").ToList();
+                    if (saved_svgs != null)
+                    {
+                        foreach (var saved_svg in saved_svgs)
+                        {
+                            if (saved_svg.SvgfileName == "")
+                            {
+                                _context.WarpedSvgs.Remove(saved_svg);
+                                _context.SaveChanges();
+                            }
+                        }
+                        saved_svgs = _context.WarpedSvgs.Where(s => s.UserId == "41ae9ea6-035a-4bc6-98f9-fd758422de6d" && s.SvgfileName != "").OrderByDescending(s => s.UpdatedAt).ToList();
+                        return Json(new { status = "success", saved_svgs = saved_svgs });
+                    }
+                    return Json(new { status = "success", msg = "There is no svgs." });
+                }
             }
             catch (Exception e)
             {
